feat: deprecate nvidia (api), add 3080 add 3090
