+++ README.md
@@ -34,7 +34,7 @@ The purpose of this bot is to get an Nvidia card. It tries multiple things to do
 |:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|:---:|
 | **3070**| | | | | | | | | | | |
 | **3080** | `✔` | `✔` | `✔` | `✔` | `✔` | `✔` | `✔` | `✔` | `✔` | `✔` | `✔` |
-| **3090** | | | | | | | | | | | |
+| **3090** | | | | | | | | | `✔` | | |
 
 ## Installation and prerequisites
 
@@ -123,6 +123,7 @@ Here is a list of variables that you can use to customize your newly copied `.en
 | Newegg | `newegg`|
 | Newegg (CA) | `newegg-ca`|
 | Nvidia | `nvidia`|
+| Nvidia (API) | `nvidia-api`|
 | Office Depot | `officedepot`|
 | Zotac | `zotac`|
